move test fixtures closer to usage
