+++ engine/execution/computation/computer/computer_test.go
@@ -38,7 +38,6 @@ import (
 	"github.com/onflow/flow-go/fvm/storage/testutils"
 	"github.com/onflow/flow-go/fvm/systemcontracts"
 	"github.com/onflow/flow-go/ledger"
-	"github.com/onflow/flow-go/model/convert/fixtures"
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/module/epochs"
 	"github.com/onflow/flow-go/module/executiondatasync/execution_data"
@@ -547,7 +546,7 @@ func TestBlockExecutor_ExecuteBlock(t *testing.T) {
 		serviceEvents, err := systemcontracts.ServiceEventsForChain(execCtx.Chain.ChainID())
 		require.NoError(t, err)
 
-		payload, err := json.Decode(nil, []byte(fixtures.EpochSetupFixtureJSON))
+		payload, err := json.Decode(nil, []byte(unittest.EpochSetupFixtureJSON))
 		require.NoError(t, err)
 
 		serviceEventA, ok := payload.(cadence.Event)
@@ -558,7 +557,7 @@ func TestBlockExecutor_ExecuteBlock(t *testing.T) {
 		}
 		serviceEventA.EventType.QualifiedIdentifier = serviceEvents.EpochSetup.QualifiedIdentifier()
 
-		payload, err = json.Decode(nil, []byte(fixtures.EpochCommitFixtureJSON))
+		payload, err = json.Decode(nil, []byte(unittest.EpochCommitFixtureJSON))
 		require.NoError(t, err)
 
 		serviceEventB, ok := payload.(cadence.Event)
